feat(auth): two-step password reset with token + OTP
Wire up the three reset endpoints behind the spec:
- POST /auth/forgot-password now generates a URL-safe token + 6-digit
  OTP, stores SHA-256 hashes, applies an IP rate-limit (5/h) and a
  per-user 24h cap (3 valid tokens), and emails the link. Mock accounts
  in DEMO_MODE still get the seed password back in the response.
- POST /auth/reset/verify-otp validates the OTP, increments
  otp_attempts on failure, and locks the row at 5 strikes.
- POST /auth/reset/confirm re-validates token+OTP (defence in depth),
  updates password_hash, sets used_at, and writes the
  password_reset_completed audit log.

Error responses use the project's standard {success,data,message,code}
shape via a small _err() helper, since the global HTTPException handler
flattens structured detail into a string. Bumps the password-reset PK
to Integer to match the rest of the models and avoid SQLite
autoincrement quirks on BigInteger.

13/13 e2e tests pass.

Also resolves a pre-existing merge conflict in admin/SettingsPage.tsx
by taking --ours (the main-branch version). Backup of the conflicted
state at /tmp/SettingsPage.tsx.conflict-backup if it needs to be
re-merged.

+++ backend/alembic/versions/t1u2v3w4x5y6_add_password_reset_tokens.py
@@ -20,7 +20,7 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "password_reset_tokens",
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
         sa.Column("user_id", sa.Integer(), nullable=False),
         sa.Column("token_hash", sa.String(length=64), nullable=False, unique=True),
         sa.Column("otp_hash", sa.String(length=64), nullable=False),
